backend/requirements.txt
django
psycopg2
djangorestframework
gunicorn
Pillow
newsapi-python